+++ docs-src/src/pages/premium.tsx
@@ -231,6 +231,16 @@ export default function Premium() {
                                         </div>
                                     </div>
                                 </a>
+                                <a href="/logger.html" target="_blank">
+                                    <div className="premium-block hover-shadow-middle bg-gradient-right-top">
+                                        <div className="premium-block-inner">
+                                            <h4>RxServer Koa Adapter</h4>
+                                            <p>
+                                                An adapter to use the RxServer with Koa instead of express. Used to have better performance when serving requests.
+                                            </p>
+                                        </div>
+                                    </div>
+                                </a>
                             </div>
                         </div>
                     </div>
@@ -539,6 +549,11 @@ export default function Premium() {
                                                         <li>
                                                             <a href="/rx-server.html" target="_blank">
                                                                 RxServer Adapter Fastify
+                                                            </a>
+                                                        </li>
+                                                        <li>
+                                                            <a href="/rx-server.html" target="_blank">
+                                                                RxServer Adapter Koa
                                                             </a>
                                                         </li>
                                                     </ul>